packages/opengsync-server/pyproject.toml
[build-system]
requires = ["setuptools>=70.1", "wheel>=0.43"]
build-backend = "setuptools.build_meta"

[project]
name = "opengsync-server"
version = "1.0"
description = "Web server for NGS sample/library/project tracking"
authors = [
    { name = "lutrarutra", email = "lutrarutra@pm.me" }
]
requires-python = ">=3.11"

dependencies = [
    "numpy >= 2.2",
    "pandas >= 2.3",
    "bcrypt >= 4.0.1",
    "blinker >= 1.6.2",
    "click >= 8.1.7",
    "dnspython >= 2.4.2",
    "email-validator >= 2.0.0.post2",
    "Flask >= 3.0.3",
    "Flask-Bcrypt >= 1.0.1",
    "flask-htmx >= 0.3.2",
    "Flask-Login >= 0.6.3",
    "Flask-Mail >= 0.10.0",
    "Flask-WTF >= 1.2.1",
    "flask-session >= 0.8.0",
    "flask-caching >= 2.3.0",
    "pygments",
    "greenlet >= 2.0.2",
    "idna >= 3.4",
    "premailer >= 3.10.0",
    "iniconfig >= 2.0.0",
    "redis",
    "itsdangerous >= 2.1.2",
    "Jinja2 >= 3.1.2",
    "loguru >= 0.7.2",
    "MarkupSafe >= 2.1.3",
    "pluggy >= 1.3.0",
    "pydantic >= 1.10.12",
    "python-dateutil >= 2.8.2",
    "python-dotenv >= 1.0.0",
    "pytz >= 2023.3.post1",
    "six >= 1.16.0",
    "typing_extensions >= 4.8.0",
    "tzdata >= 2023.3",
    "WTForms >= 3.0.1",
    "gunicorn >= 21.2.0",
    "et-xmlfile >= 1.1.0",
    "openpyxl >= 3.1.2",
    "cryptography >= 42.0.5",
    "pyopenssl >= 24.0.0",
    "patsy >= 0.5.6",
    "plotly >= 5.19.0",
    "plotly-express >= 0.4.1",
    "scipy >= 1.12.0",
    "statsmodels >= 0.14.1",
    "tenacity >= 8.2.3",
    "pyyaml >= 6.0.1",
    "google-genai >= 1.28",
    "flask-limiter >= 3.12",
    "framehints",
]

[tool.flake8]
max-line-length = 88
exclude = [
    ".git",
    "__pycache__",
]
extend-ignore = ["E501", "W292", "W293"]

[project.urls]
Repository = "https://github.com/lutrarutra/opengsync-app"


